fix: move path constants to avoid bloating lti and iframe bundles
